Finish user integration with fire base back end

+++ lib/Core/NetWork/storage_services.dart
@@ -1,11 +1,10 @@
 import 'dart:io';
 import 'package:firebase_storage/firebase_storage.dart';
+import 'package:intl/intl.dart';
 import 'package:path/path.dart';
 
 class StorageService {
-  final FirebaseStorage _firebaseStorage;
-
-  StorageService(this._firebaseStorage);
+  final FirebaseStorage _firebaseStorage = FirebaseStorage.instance;
 
   Future<String?> uploadPhoto({
     required File photo,
@@ -13,7 +12,9 @@ class StorageService {
   }) async {
     try {
       String photoName = basename(photo.path);
-      Reference ref = _firebaseStorage.ref(folderName).child(photoName);
+      String timeStamp = DateFormat('yyyyMMddHHmmss').format(DateTime.now());
+      Reference ref =
+          _firebaseStorage.ref(folderName).child('$photoName $timeStamp');
       await ref.putFile(photo);
       return await ref.getDownloadURL();
     } catch (e) {
@@ -21,7 +22,10 @@ class StorageService {
     }
   }
 
-  Future<void> deletePhoto({required String url}) async {
+  Future<void> deletePhoto({required String? url}) async {
+    if (url == null) {
+      return;
+    }
     await _firebaseStorage.refFromURL(url).delete();
   }
 }
